Check moving in all 4 hemispheres
This is finding two different bugs that will need to be fixed. First off,
western longitudes and southern latitudes are shown as negative in
addition to the direction marker changing making it a bit of a double
negative.

Second, the longitude marker for East/West appears to be backwards. Western
longitudes show as negative east, and eastern longitudes as positive west.