Added code for directly testing the back-end formatting of coordinates
This code is finding the same bug already detected by the MapModeTests
albeit in a much more direct way without needing to instantiate the
user interface. Ideally, this should be a unit test that can be executed
on the host-side, but the code depends on the Location.convert() for
the actual coordinate conversions which is only available on the
Android platform itself.