Converted all code added in last commit from Kotlin to Java
The Kotlin code was causing build errors on every file in the
compileDebugScala task. It appears like it's trying to compile it as
Scala or something. Avoiding Kotlin until this can be resolved.

Fixed up branch to rebase it off of master

Reverted Kotlin plugin changes

As Kotlin can't be used currently, leave out the changes to support it.