digirig: Basic implementation of AfskProto and Digirig
This includes the changes that I needed to make to Charlie's original
commit, as well as the changes I made to the copied KissProto
implementation.  The UI stuff on the app looks mostly correct at this
point, and I believe the bones are there to get packet tx/rx working.