backed/DigiRig: Sync RTS signal to audio playback
I originally had some hacky code in here to _hopfeully_ ensure that the
RTS signal was raise while packets were being transmitted, but it was
only a best-guess attempt.  This commit implements an
OnPlaybackPositionUpdateListener to tell us exactly when the audio for a
packet has finished playing, so we know exactly when to clear the RTS
signal.