Attempt to bump Java version for Emulator
The latest SDK Commandline tools starting with release 9.0 require Java
17 to execute. These are used by the Android Emulator GitHub Action so
it will fail to run if the default Java environment is older than 17.
Need JRE 17 to be installed and default, but not the full JDK. However,
the build toolchain for Scala still requires to be run under Java 11 so
we need both installed. We just need to make sure JAVA_HOME is set
appropriately for JDK 11 when calling Gradle.

Set JAVA_HOME to JDK 11 for gradle on instrumented tests.